Add files via upload
First Part of the project: From the very large set of apps in the health & fitness and medical categories, we will develop automated methods to extract information from the apps’ webpages on the US Apple app store and we will build a database of apps.

Second Part of the project: From the unstructured text related to apps collected in our database, we will develop automated methods to characterize apps, specifically to classify apps into the different medical specialties (e.g. cardiology, audiology, oncology, …).